pyproject.toml
[tool.poetry]
name = "repl-nix-peppygifs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
imagehash = "^4.3.1"
pillow = "^10.4.0"
boto3 = "^1.35.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
